Refactor documents app, add imports app
