+++ apps/documents/views.py
@@ -9,18 +9,15 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 from influxdb_metrics.loader import log_metric
 
-from apps.authentication import DocsLambdaRequest, get_jwt_from_request
+from apps.authentication import DocsLambdaRequest
 from apps.jobs.models import AsyncActionType
 from apps.permissions import get_owner_id
-from .filters import DocumentFilterSet, CsvDocumentFilterSet
-from .models import Document, CsvDocument, UploadStatus
-from .permissions import UserHasPermission, UserHasCsvFilePermission
+from apps.utils import UploadStatus
+from .filters import DocumentFilterSet
+from .models import Document
+from .permissions import UserHasPermission
 from .rpc import DocumentRPCClient
-from .serializers import (ShipmentDocumentSerializer,
-                          ShipmentDocumentCreateSerializer,
-                          CsvDocumentSerializer,
-                          CsvDocumentCreateSerializer,
-                          CsvDocumentCreateResponseSerializer, )
+from .serializers import ShipmentDocumentSerializer, ShipmentDocumentCreateSerializer
 
 LOG = logging.getLogger('transmission')
 
@@ -73,54 +70,13 @@ class DocumentViewSet(mixins.CreateModelMixin,
         LOG.debug(f'Updating document {instance.id} with new details.')
         log_metric('transmission.info', tags={'method': 'documents.update', 'module': __name__})
 
-        serializer = self.serializer_class(instance, data=request.data, partial=partial, context={'auth': request.auth})
+        serializer = self.get_serializer_class()(instance, data=request.data, partial=partial,
+                                                 context={'auth': request.auth})
         serializer.is_valid(raise_exception=True)
 
         self.perform_update(serializer)
 
         return Response(serializer.data, status=status.HTTP_200_OK)
-
-
-class CsvDocumentViewSet(mixins.CreateModelMixin,
-                         mixins.RetrieveModelMixin,
-                         mixins.UpdateModelMixin,
-                         mixins.ListModelMixin,
-                         viewsets.GenericViewSet):
-    queryset = CsvDocument.objects.all()
-    serializer_class = CsvDocumentSerializer
-    permission_classes = (permissions.IsAuthenticated, UserHasCsvFilePermission, )
-    filter_backends = (filters.DjangoFilterBackend, )
-    filter_class = CsvDocumentFilterSet
-
-    def get_queryset(self):
-        return self.queryset.filter(updated_by=self.request.user.id)
-
-    def perform_create(self, serializer):
-        if settings.PROFILES_ENABLED:
-            created = serializer.save(owner_id=get_owner_id(self.request), updated_by=self.request.user.id)
-        else:
-            created = serializer.save()
-        return created
-
-    def perform_update(self, serializer):
-        if settings.PROFILES_ENABLED:
-            updated = serializer.save(updated_by=self.request.user.id)
-        else:
-            updated = serializer.save()
-        return updated
-
-    def create(self, request, *args, **kwargs):
-        """
-        Create a pre-signed s3 post and create a corresponding document object with pending status
-        """
-        LOG.debug(f'Creating a CSV document object.')
-        log_metric('transmission.info', tags={'method': 'documents.create', 'module': __name__})
-
-        serializer = CsvDocumentCreateSerializer(data=request.data, context={'auth': get_jwt_from_request(request)})
-        serializer.is_valid(raise_exception=True)
-        doc_obj = self.perform_create(serializer)
-
-        return Response(CsvDocumentCreateResponseSerializer(doc_obj).data, status=status.HTTP_201_CREATED)
 
 
 class S3Events(APIView):
